Prueba cypress: Elimina un ingreso de la lista de ingresos solo si confirmas

+++ src/presenter.js
@@ -388,9 +388,12 @@ function actualizarListaIngreso(ingreso){
 }
 
 function eliminarIngreso(index) {
-  lista_ingresos.eliminarIngreso(index);
-  actualizarListaIngreso(); 
-  div_ingreso.innerHTML = "<p>Ingreso eliminado</p>";
+  const confirmacion = confirm("¿Estás seguro de que deseas eliminar este ingreso?");
+  if (confirmacion) {
+    lista_ingresos.eliminarIngreso(index);
+    actualizarListaIngreso(); 
+    div_ingreso.innerHTML = "<p>Ingreso eliminado</p>";
+  }
 }
 
 function seleccionarIngreso(index) {
